[R] Major command framework integration.
