+++ pom.xml
@@ -20,6 +20,11 @@
             <id>placeholderapi</id>
             <url>http://repo.extendedclip.com/content/repositories/placeholderapi/</url>
         </repository>
+
+        <repository>
+            <id>jitpack.io</id>
+            <url>https://jitpack.io</url>
+        </repository>
     </repositories>
     <dependencies>
         <!--Spigot API and NMS-->
@@ -57,6 +62,12 @@
             <groupId>com.github.cryptomorin</groupId>
             <artifactId>XSeries</artifactId>
             <version>8.2.0</version>
+        </dependency>
+
+        <dependency>
+            <groupId>com.github.MartenM</groupId>
+            <artifactId>SimpleCommands</artifactId>
+            <version>1.3.0</version>
         </dependency>
 
     </dependencies>
