+++ src/main/java/nl/martenm/servertutorialplus/commands/ServerTutorialRootCommand.java
@@ -1,0 +1,33 @@
+package nl.martenm.servertutorialplus.commands;
+
+import nl.martenm.servertutorialplus.commands.help.STFormatter;
+import nl.martenm.servertutorialplus.commands.sub.*;
+import nl.martenm.servertutorialplus.commands.sub.manage.ReloadPluginCommand;
+import nl.martenm.servertutorialplus.commands.sub.manage.SaveCommand;
+import nl.martenm.servertutorialplus.commands.sub.points.*;
+import nl.martenm.servertutorialplus.commands.sub.tutorials.*;
+import nl.martenm.simplecommands.RootCommand;
+
+public class ServerTutorialRootCommand extends RootCommand {
+
+    public ServerTutorialRootCommand() {
+        super("servertutorial", "servertutorialplus.commands", false);
+        setHelpFormatter(new STFormatter());
+
+        addCommand(new CreateTutorialCommand());
+        addCommand(new RemoveTutorialCommand());
+        addCommand(new AddpointCommand());
+        addCommand(new RemovePointCommand());
+        addCommand(new EditTutorialCommand());
+        addCommand(new EditPointCommand());
+        addCommand(new EditAllPointsCommand());
+        addCommand(new InfoCommand());
+        addCommand(new PlayerCommand());
+        addCommand(new QuitTutorialCommand());
+        addCommand(new PlayPointCommand());
+        addCommand(new SetBlockCommand());
+        addCommand(new NpcCommand());
+        addCommand(new ReloadPluginCommand());
+        addCommand(new SaveCommand());
+    }
+}
